Modify liminal sdk event manager to use a custom data class for events
replaced action queue for a list of custom class 'ActionEventData' which houses the action as well as additional properties. Added additional methods for calling individual actions as well as multiple actions at once. Converted methods into IEnumerators to allow for delayed actions and to make it more scalable in future.