Implement liminal sdk event manager (#52)
* Implement liminal sdk event manager

Implement a static script to allow actions to be queued on both platform and sdk side that can be then triggered by either side when required.

* Modify liminal sdk event manager to use a custom data class for events

replaced action queue for a list of custom class 'ActionEventData' which houses the action as well as additional properties. Added additional methods for calling individual actions as well as multiple actions at once. Converted methods into IEnumerators to allow for delayed actions and to make it more scalable in future.

* Update QueueEvent method to AddEvent in sdk event manager

Updated the QueveEvent method to AddEvent to be more inline with how the system currently works.

* Add sdk event manager example and adjust event manager to allow for events to be kept after use, and remove event methods

Added an example script of how to use the sdk event manager to register and trigger events. Updated the event manager to have functionality to keep events after they have been invoked for repeat use if necessary.  Added methods to remove action event data by id, allowing events to be removed when required.

* Clear up typo in sdk event manager example script

* Update liminal sdk event manager to separate out event data class, rename methods, change namespace

Made adjustments to the sdk event manager to move event data class into its own class rather than be nested, renamed coroutine methods to include the 'routine' convention in the naming, and adjusted the namespace to something more appropriate.